triangle function to modify with float

+++ srcs/quaternion/q_rotate.c
@@ -17,11 +17,11 @@ t_quaternion	rotate_q(t_vec4 *n, float theta)
 	t_quaternion		q;
 	float				s;
 
-	s = sin(theta * 0.5);
+	s = sinf(theta * 0.5);
 	q.x = s * n->arr[0];
 	q.y = s * n->arr[1];
 	q.z = s * n->arr[2];
-	q.w = cos(theta * 0.5);
+	q.w = cosf(theta * 0.5);
 	return (q);
 }
 
@@ -29,10 +29,10 @@ t_vec4			rotate(t_quaternion *q, t_vec4 *v, t_quaternion *q_i)
 {
 	t_quaternion	res;
 
-	v->arr[3] = 0;
+	v->arr[3] = 0.0f;
 	res = q_mul_q(q, (t_quaternion *)v);
 	res = q_mul_q(&res, q_i);
-	v->arr[3] = 1;
-	res.w = 1;
+	v->arr[3] = 1.0f;
+	res.w = 1.0f;
 	return (*((t_vec4 *)&res));
 }
